feat(builder):couple builders and add dockerbuilder pod
