--- bpbuilder/deis-slugbuilder.yaml
@@ -1,29 +1,0 @@
-# A debugging utility for testing Minio from within k8s.
-apiVersion: v1
-kind: Pod
-metadata:
-  name: deis-slugbuilder
-  labels:
-    heritage: deis
-    version: 2015-sept
-spec:
-  restartPolicy: Never
-  containers:
-    - name: deis-slugbuilder
-      imagePullPolicy: Always
-      image: smothiki/slugbuilder:5.3
-      env:
-        - name: DEBUG
-          value: "1"
-        - name: TAR_URL
-          value: tar-url
-        - name : put-url
-          value: puturl
-      volumeMounts:
-        - name: object-store
-          mountPath: /var/run/secrets/object/store
-          readOnly: true
-  volumes:
-    - name: object-store
-      secret:
-        secretName: object-store
