--- bpbuilder/entrypoint.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-set -e
-
-bpbuilder & 
-exec fetcher
