--- manifests/deis-bpbuilder-rc.yaml
@@ -1,26 +1,0 @@
-apiVersion: v1
-kind: ReplicationController
-metadata:
-  # Change the name
-  name: deis-builderbp-rc
-  labels:
-    heritage: deis
-spec:
-  replicas: 1
-
-  # Change the selector name
-  selector:
-    app: deis-builder
-  template:
-    metadata:
-      labels:
-        # Important: these labels need to match the selector above
-        app: deis-builder
-    spec:
-      containers:
-        - name: deis-bpbuilder
-        image: deis/bpbuilder:2.0.0
-          ports:
-            - containerPort: 2222
-        command:
-          - entrypoint.sh
