+++ manifests/deis-builder-rc.yaml
@@ -1,0 +1,36 @@
+apiVersion: v1
+kind: ReplicationController
+metadata:
+  # Change the name
+  name: deis-builder
+  labels:
+    heritage: deis
+spec:
+  replicas: 1
+
+  # Change the selector name
+  selector:
+    app: deis-builder
+  template:
+    metadata:
+      labels:
+        # Important: these labels need to match the selector above
+        app: deis-builder
+    spec:
+      containers:
+        - name: deis-builder
+          image: smothiki/builder:2.5.2
+          ports:
+            - containerPort: 2223
+            - containerPort: 3000
+          env:
+            - name: "EXTERNAL_PORT"
+              value: "2223"
+  #     volumeMounts:
+  #       - name: object-store
+  #         mountPath: /var/run/secrets/object/store
+  #         readOnly: true
+  # volumes:
+  #   - name: object-store
+  #     secret:
+  #       secretName: object-store
