+++ manifests/deis-builder-service.yaml
@@ -9,11 +9,7 @@ metadata:
 spec:
   ports:
     - port: 2222
-      targetPort: 2222
-      name: ssh
-      protocol: TCP
-    - port: 2223
-      targetPort: 2222
+      targetPort: 2223
       name: ssh
       protocol: TCP
   selector:
