--- manifests/dfbuilder/deis-example-service.yaml
@@ -1,17 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  # Give this a useful name.
-  name: deis-builder
-  labels:
-    heritage: deis
-    release: 2.0.0
-spec:
-  ports:
-    - port: 2223
-      targetPort: 2223
-      name: ssh
-      protocol: TCP
-  selector:
-    # Match the selector in the RC
-    app: deis-bpbuilder
