+++ rootfs/check-repos
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+export ETCD=$DEIS_ETCD_1_SERVICE_HOST:$DEIS_ETCD_1_SERVICE_PORT_CLIENT
+
+cd $(dirname $0) # absolute path
+
+for repo in $(ls | grep .git)
+do
+    reponame="${repo%.*}"
+    etcdctl -C "$ETCD" ls /deis/services/"$reponame" > /dev/null 2>&1
+    if [[ $? -eq 4 ]]
+    then
+        rm -rf "$repo"
+        appname="$DEIS_REGISTRY_SERVICE_HOST:$DEIS_REGISTRY_SERVICE_PORT/$reponame"
+        docker images | grep $appname | awk '{ print $3 }' | xargs -r docker rmi -f
+        # remove any dangling images left over from the cleanup
+        docker images --filter "dangling=true" | awk '{ print $3 }' | grep -v IMAGE | xargs -r docker rmi -f
+    fi
+done
