+++ rootfs/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+fetcher &
+exec builder
