--- rootfs/etc/confd/conf.d/check-repos.toml
@@ -1,9 +1,0 @@
-[template]
-src   = "check-repos"
-dest  = "/home/git/check-repos"
-uid = 0
-gid = 0
-mode  = "0755"
-keys = [
-  "/deis/registry",
-]
