--- rootfs/etc/confd/conf.d/push-images.toml
@@ -1,9 +1,0 @@
-[template]
-src   = "push-images"
-dest  = "/usr/local/bin/push-images"
-uid = 0
-gid = 0
-mode  = "0755"
-keys = [
-  "/deis/registry",
-]
