+++ rootfs/etc/confd/templates/builder
@@ -1,0 +1,161 @@
+#!/usr/bin/env bash
+#
+# builder hook called on every git receive-pack
+# NOTE: this script must be run as root (for docker access)
+#
+set -eo pipefail
+
+ARGS=3
+HOST=`ifconfig eth0 | grep 'inet addr:' | cut -d: -f2 | awk '{ print $1}'`
+indent() {
+    echo "       $@"
+}
+
+puts-step() {
+    echo "-----> $@"
+}
+
+puts-step-sameline() {
+    echo -n "-----> $@"
+}
+
+puts-warn() {
+    echo " !     $@"
+}
+
+usage() {
+    echo "Usage: $0 <user> <repo> <sha>"
+}
+
+parse-string(){
+    # helper to avoid the single quote escape
+    # occurred in command substitution
+    local args=() idx=0 IFS=' ' c
+    for c; do printf -v args[idx++] '%s ' "$c"; done
+    printf "%s\n" "${args[*]}"
+}
+
+if [ $# -ne $ARGS ]; then
+    usage
+    exit 1
+fi
+
+USER=$1
+REPO=$2
+GIT_SHA=$3
+SHORT_SHA=${GIT_SHA:0:8}
+APP_NAME="${REPO%.*}"
+
+cd $(dirname $0) # ensure we are in the root dir
+
+ROOT_DIR=$(pwd)
+REPO_DIR="${ROOT_DIR}/${REPO}"
+BUILD_DIR="${REPO_DIR}/build"
+CACHE_DIR="${REPO_DIR}/cache"
+
+# define image names
+SLUG_NAME="$APP_NAME:git-$SHORT_SHA"
+META_NAME=`echo ${SLUG_NAME}| tr ":" "-"`
+TMP_IMAGE="$DEIS_REGISTRY_SERVICE_HOST:$DEIS_REGISTRY_SERVICE_PORT/$IMAGE_NAME"
+# create app directories
+mkdir -p $BUILD_DIR $CACHE_DIR
+# create temporary directory inside the build dir for this push
+TMP_DIR=$(mktemp -d -p $BUILD_DIR)
+
+cd $REPO_DIR
+# use Procfile if provided, otherwise try default process types from ./release
+USING_DOCKERFILE=true
+if [ -f Procfile ]; then
+    PROCFILE=$(cat Procfile | yaml2json-procfile)
+    USING_DOCKERFILE=false
+    POD="slugbuilder"
+else
+    PROCFILE="{}"
+fi
+
+if [[ ! -f /var/run/secrets/object/store/access-key-id  ]]; then
+  if [ -f Dockerfile ]; then
+    l1=`grep -n "object-store" /etc/deis-dockerbuilder.yaml | head -n1 |cut -d ":" -f1`
+    l2=$(($l1+3))
+    sed "$l1,$l2 d" /etc/deis-dockerbuilder.yaml > /etc/${SLUG_NAME}.yaml.tmp
+    l1=`grep -n "object-store" /etc/deis-dockerbuilder.yaml.tmp | head -n1 |cut -d ":" -f1`
+    l2=$(($l1+3))
+    sed "$l1,$l2 d" /etc/${SLUG_NAME}.yaml.tmp > /etc/${SLUG_NAME}.yaml
+    sed -i -- "s#repo_name#$TMP_IMAGE#g" /etc/${SLUG_NAME}.yaml
+  else
+    head -n 21 /etc/deis-slugbuilder.yaml > /etc/${SLUG_NAME}.yaml
+  fi
+else
+  if [ -f Dockerfile ]; then
+    cp /etc/deis-dockerbuilder.yaml /etc/${SLUG_NAME}.yaml
+    sed -i -- "s#repo_name#$TMP_IMAGE#g" /etc/${SLUG_NAME}.yaml
+  else
+    cp /etc/deis-slugbuilder.yaml /etc/${SLUG_NAME}.yaml
+  fi
+fi
+
+git archive --format=tar.gz ${GIT_SHA} > ${APP_NAME}.tar.gz
+
+
+# if minio is in the cluster, use it. otherwise use fetcher
+if [[ -n "$DEIS_MINIO_SERVICE_HOST" && -n "$DEIS_MINIO_SERVICE_PORT" ]]; then
+  S3EP=${DEIS_MINIO_SERVICE_HOST}:${DEIS_MINIO_SERVICE_PORT}
+elif [ -z "$S3EP" ]; then
+  S3EP=${HOST}:3000
+fi
+
+TAR_URL=http://$S3EP/git/home/${SLUG_NAME}/tar
+PUSH_URL=http://$S3EP/git/home/${SLUG_NAME}/push
+
+sed -i -- "s#repo_name#$META_NAME#g" /etc/${SLUG_NAME}.yaml
+sed -i -- "s#puturl#$PUSH_URL#g" /etc/${SLUG_NAME}.yaml
+sed -i -- "s#tar-url#$TAR_URL#g" /etc/${SLUG_NAME}.yaml
+
+kubectl create -f /etc/${SLUG_NAME}.yaml
+
+#check for image creation or slug existence in S3EP
+
+while [ ! -f /apps/${SLUG_NAME}/slug.tgz ]
+do
+  sleep 2
+  echo "no file"
+done
+
+URL="http://$DEIS_WORKFLOW_SERVICE_HOST:$DEIS_WORKFLOW_SERVICE_PORT/v1/hooks/config"
+RESPONSE=$(get-app-config -url="$URL" -key="{{ getv "/deis/controller/builderKey" }}" -user=$USER -app=$APP_NAME)
+CODE=$?
+echo $URL
+echo $RESPONSE
+if [ $CODE -ne 0 ]; then
+    puts-warn $RESPONSE
+    exit 1
+fi
+
+# use Procfile if provided, otherwise try default process types from ./release
+echo $PROCFILE
+
+puts-step "Launching... "
+URL="http://$DEIS_WORKFLOW_SERVICE_HOST:$DEIS_WORKFLOW_SERVICE_PORT/v1/hooks/build"
+DATA="$(generate-buildhook "$SHORT_SHA" "$USER" "$APP_NAME" "$APP_NAME" "$PROCFILE" "$USING_DOCKERFILE")"
+echo $DATA
+PUBLISH_RELEASE=$(echo "$DATA" | publish-release-controller -url=$URL -key={{ getv "/deis/controller/builderKey" }})
+echo $PUBLISH_RELEASE
+CODE=$?
+if [ $CODE -ne 0 ]; then
+    puts-warn "ERROR: Failed to launch container"
+    puts-warn $PUBLISH_RELEASE
+    exit 1
+fi
+
+RELEASE=$(echo $PUBLISH_RELEASE | extract-version)
+DOMAIN=$(echo $PUBLISH_RELEASE | extract-domain)
+indent "done, $APP_NAME:v$RELEASE deployed to Deis"
+echo
+indent "http://$DOMAIN"
+echo
+indent "To learn more, use \`deis help\` or visit http://deis.io"
+echo
+
+# cleanup
+cd $REPO_DIR
+git gc &>/dev/null
