+++ rootfs/etc/confd/templates/receiver
@@ -12,4 +12,4 @@ curl \
   -H 'Content-Type: application/json' \
   -H "X-Deis-Builder-Auth: {{ getv "/deis/controller/builderKey" }}" \
   -d "{\"receive_user\": \"$username\", \"receive_repo\": \"$app\", \"sha\": \"$sha\", \"fingerprint\": \"$fingerprint\", \"ssh_connection\": \"$SSH_CONNECTION\", \"ssh_original_command\": \"$SSH_ORIGINAL_COMMAND\"}" \
-  --silent {{ getv "/deis/controller/protocol" }}://{{ getv "/deis/controller/host" }}:{{ getv "/deis/controller/port" }}/v1/hooks/push >/dev/null
+  --silent "http://$DEIS_WORKFLOW_SERVICE_HOST:$DEIS_WORKFLOW_SERVICE_PORT/v1/hooks/push" >/dev/null
