+++ rootfs/etc/deis-slugbuilder.yaml
@@ -2,22 +2,22 @@
 apiVersion: v1
 kind: Pod
 metadata:
-  name: deis-slugbuilder
+  name: repo_name
   labels:
     heritage: deis
     version: 2015-sept
 spec:
   restartPolicy: Never
   containers:
-    - name: deis-slugbuilder
+    - name: repo_name
       imagePullPolicy: Always
-      image: smothiki/slugbuilder:5.3
+      image: smothiki/slugbuilder:5.5
       env:
         - name: DEBUG
           value: "1"
         - name: TAR_URL
           value: tar-url
-        - name : put-url
+        - name : put_url
           value: puturl
       volumeMounts:
         - name: object-store
